Move activity edit dialogs and simplify props
(cherry picked from commit e726a7f513817fcf53ddadb86b326fa8f90f0877)
