+++ frontend/src/components/activity/dialog/FormScheduleEntryItem.vue
