+++ frontend/src/locales/it.json
@@ -23,6 +23,14 @@
       "copyActivityInfoDialog": {
         "allow": "Consenti ora"
       },
+      "dialog": {
+        "dialogActivityEdit": {
+          "title": "Modificare l'actività"
+        },
+        "formScheduleEntryList": {
+          "name": "Programmato"
+        }
+      },
       "menuCardlessContentNode": {
         "deletingDisabled": "Deve essere vuoto per essere eliminato"
       },
