+++ frontend/src/locales/rm.json
@@ -17,6 +17,14 @@
           }
         }
       },
+      "dialog": {
+        "dialogActivityEdit": {
+          "title": "Modifitgar l'activitad"
+        },
+        "formScheduleEntryList": {
+          "name": "Termins planisads"
+        }
+      },
       "menuCardlessContentNode": {
         "deletingDisabled": "Sto esser vid per stizzar"
       },
